+++ src/main/java/PrograIV/bolsaEmpleo/data/CaracteristicaRepository.java
@@ -1,4 +1,5 @@
 package PrograIV.bolsaEmpleo.data;
+
 import PrograIV.bolsaEmpleo.Logic.Caracteristica;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -6,10 +7,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CaracteristicaRepository extends JpaRepository<Caracteristica, Integer> {
-    // Raíces del árbol jerárquico (sin padre)
+public interface CaracteristicaRepository extends JpaRepository<Caracteristica, Long> {
+
+    // Raíces del árbol jerárquico (las que no tienen padre)
     List<Caracteristica> findByPadreIsNull();
 
     // Hijos directos de una característica padre
-    List<Caracteristica> findByPadreId(Integer padreId);
+    List<Caracteristica> findByPadreId(Long padreId);
 }
